Add files via upload
SYSTEM PACKAGES
sys : provides access to variables used or maintained by the interpreter and to functions that interact strongly with the interpreter. 
os: provides a portable way of using operating system dependent functionality. 
subprocess: spawns new processes, connects to their input/output/error pipes, and obtains their return codes.
smtplib: defines an SMTP client session object that can be used to send mail to any Internet machine with an SMTP or ESMTP listener daemon.
logging: defines functions and classes which implement a flexible event logging system for applications and libraries.
SPECIALIZED PACKAGES
Pandas: an open source, BSD-licensed library providing high-performance, easy-to-use data structures and data analysis tools for the Python programming language. This is the primary package we use for data management and analysis. It can be used to read and manipulate SAS files, as well as other data types. It is used in this script to read and subset a SAS metadata data set and a format data set (created from a format catalog) to determine which studies to run and to determine file names.
SASPy: a Python module written by SAS to provide Python APIs to the SAS system, enabling users to start a SAS session and run SAS code from Python.[7] It is used in this script to read a SAS format catalog, then use the cntlout option to write a SAS data set from the format procedure.